refactor(board): move state timers to hooks/useGameTimer
